Accident_detection_2/Dashboard/fastapi-ml/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
ultralytics>=8.2.0
Pillow>=10.3.0
numpy>=1.26.0
python-multipart>=0.0.9
pydantic>=2.7.0
